+++ frontend/src/components/pages/consumers/Group.Details.tsx
@@ -53,8 +53,6 @@ class GroupDetails extends PageComponent<{ groupId: string }> {
         const requiredTopics = group.members.flatMap(m => m.assignments.map(a => a.topicName)).distinct();
 
         const totalPartitions = group.members.flatMap(m => m.assignments).sum(a => a.partitionIds.length);
-        const partitionsWithOffset = group.lag.topicLags.sum(tl => tl.partitionsWithOffset);
-        const topicsWithOffset = group.lag.topicLags.length;
 
         return (
             <MotionDiv style={{ margin: '0 1rem' }}>
@@ -67,9 +65,10 @@ class GroupDetails extends PageComponent<{ groupId: string }> {
                             <ProtocolType group={group} />
                             <Statistic title='Members' value={group.members.length} />
                             <Statistic title='Assigned Topics' value={requiredTopics.length} />
-                            <Statistic title='Topics with offset' value={topicsWithOffset} />
                             <Statistic title='Assigned Partitions' value={totalPartitions} />
-                            <Statistic title='Partitions with offset' value={partitionsWithOffset} />
+                            <Statistic title='Protocol Type' value={group.protocolType} />
+                            <Statistic title='Protocol' value={group.protocol} />
+                            <Statistic title='Coordinator ID' value={group.coordinatorId} />
                             <Statistic title='Total Lag' value={group.lagSum} />
                         </Row>
                     </Card>