+++ frontend/src/state/restInterfaces.ts
@@ -189,6 +189,7 @@ export type GroupAction = 'all' | typeof GroupActions[number];
 export interface GroupDescription {
     groupId: string; // name of the group
     state: string; // Dead, Initializing, Rebalancing, Stable
+    protocol: string;
     protocolType: string; // Will be "consumer" if we can decode the members; otherwise ".members" will be empty, which happens for "sr" (for schema registry) for example
     members: GroupMemberDescription[]; // members (consumers) that are currently present in the group
     coordinatorId: number;